feat: add metric filter from sidebar
