+++ packages/common/package.json
@@ -17,7 +17,8 @@
     "dependencies": {
         "@casl/ability": "^5.4.3",
         "@typescript-eslint/eslint-plugin": "^4.4.1",
-        "moment": "^2.29.1"
+        "moment": "^2.29.1",
+        "uuid": "^8.3.2"
     },
     "scripts": {
         "dev": "nodemon src/index.ts",
