+++ packages/frontend/src/components/TableTree.tsx
@@ -19,7 +19,7 @@ import {
     Dimension,
     fieldId,
     FieldId,
-    FieldType,
+    isFilterableField,
     Metric,
     Source,
 } from 'common';
@@ -122,16 +122,8 @@ const NodeItemButtons: FC<{
     onOpenSourceDialog: (source: Source) => void;
     isHovered: boolean;
 }> = ({ node, onOpenSourceDialog, isHovered }) => {
-    const {
-        isFilteredField,
-        isFilterableDimension,
-        addDefaultFilterForDimension,
-    } = useFilters();
+    const { isFilteredField, addFilter } = useFilters();
     const isFiltered = isFilteredField(node);
-    const onFilter =
-        node.fieldType === FieldType.DIMENSION
-            ? () => addDefaultFilterForDimension(node)
-            : () => {};
 
     const menuItems: ReactNode[] = [];
     if (node.source) {
@@ -150,15 +142,15 @@ const NodeItemButtons: FC<{
             />,
         );
     }
-    if (node.fieldType === FieldType.DIMENSION && isFilterableDimension(node)) {
+    if (isFilterableField(node)) {
         menuItems.push(
             <MenuItem
                 key="filter"
                 icon="filter"
                 text="Add filter"
                 onClick={(e) => {
-                    onFilter();
                     e.stopPropagation();
+                    addFilter(node);
                 }}
             />,
         );
